Split debundle e2e suite by concern; auto-derive temp-dir prefix
Drops the per-test `prefix` argument: runLogicalModulesE2eFixture and
expectLogicalModulesE2eRejection now take the test context `t` and
slugify `t.name` into the /tmp prefix. One less thing for each test to
hand-maintain.

Splits the single e2e test target into four focused targets so a
focused run-on-change picks up only the relevant file:

- debundle_e2e_test          - behavior preservation + plain/init wrapper
                               + collision rejection (8 tests)
- debundle_cross_module_e2e_test - closure, shared bootstrap, renamed
                                   deps (3 tests)
- debundle_naturalization_e2e_test - readable rename heuristics (3 tests)
- debundle_url_rebase_e2e_test - worker + dynamic import URL rebasing (2)

logicalModule() helper moved into pipeline_e2e_support.mjs so all four
files share it.

BAZEL_TEST_INVOCATIONS=buildbuddy:a46d98e5-8844-4a46-b9d2-12b42b2c6c2d
