+++ devinfra/js/debundle/BUILD.bazel
@@ -2,16 +2,41 @@ load("@aspect_rules_js//js:defs.bzl", "js_test")
 
 package(default_visibility = ["//visibility:public"])
 
+# Shared by every e2e test: the run_transform binary plus the harness module.
+_E2E_DATA = [
+    "pipeline_e2e_support.mjs",
+    "//devinfra/js/debundle/test_support:black_box_lib",
+    "//devinfra/js/debundle/transforms:run_transform",
+]
+
+_E2E_ENV = {
+    "DUCKTAPE_RUN_TRANSFORM_BIN": "$(rootpath //devinfra/js/debundle/transforms:run_transform)",
+}
+
 js_test(
     name = "debundle_e2e_test",
-    data = [
-        "debundle_e2e_test.mjs",
-        "pipeline_e2e_support.mjs",
-        "//devinfra/js/debundle/test_support:black_box_lib",
-        "//devinfra/js/debundle/transforms:run_transform",
-    ],
+    data = _E2E_DATA + ["debundle_e2e_test.mjs"],
     entry_point = "debundle_e2e_test.mjs",
-    env = {
-        "DUCKTAPE_RUN_TRANSFORM_BIN": "$(rootpath //devinfra/js/debundle/transforms:run_transform)",
-    },
+    env = _E2E_ENV,
+)
+
+js_test(
+    name = "debundle_cross_module_e2e_test",
+    data = _E2E_DATA + ["debundle_cross_module_e2e_test.mjs"],
+    entry_point = "debundle_cross_module_e2e_test.mjs",
+    env = _E2E_ENV,
+)
+
+js_test(
+    name = "debundle_naturalization_e2e_test",
+    data = _E2E_DATA + ["debundle_naturalization_e2e_test.mjs"],
+    entry_point = "debundle_naturalization_e2e_test.mjs",
+    env = _E2E_ENV,
+)
+
+js_test(
+    name = "debundle_url_rebase_e2e_test",
+    data = _E2E_DATA + ["debundle_url_rebase_e2e_test.mjs"],
+    entry_point = "debundle_url_rebase_e2e_test.mjs",
+    env = _E2E_ENV,
 )
